Testes/Integrados/requirements.txt
pytest==7.3.1
allure-pytest==2.13.2
pytest-flask==1.2.0
requests==2.31.0
